Lift CompetitionsTables filtering to Competitions container
